myshop/requirements.txt
amqp==2.6.1
billiard==3.5.0.5
cairocffi==1.2.0
CairoSVG==2.5.0
celery==4.1.0
cffi==1.14.4
cssselect2==0.4.1
defusedxml==0.6.0
Django==2.1.5
django-admin-interface==0.14.2
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-crispy-forms==1.10.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-taggit==0.22.2
html5lib==1.1
kombu==4.6.11
pdfrw==0.4
Pillow
pycparser==2.20
Pyphen==0.10.0
pytz==2020.4
six==1.15.0
tinycss2==1.1.0
vine==1.3.0
WeasyPrint==0.42.3
webencodings==0.5.1
